requirements.txt
langchain 
langchain_community 
fastapi
uvicorn
streamlit_chat 
chromadb 
pypdf 
fastembed
python-multipart